requirements.txt
# Core web framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# PDF processing
PyMuPDF==1.23.14
pdfplumber==0.10.3

# Text-to-Speech APIs
openai==1.6.1
google-cloud-texttospeech==2.16.3

# Audio processing
pydub==0.25.1
mutagen==1.47.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Development and testing
pytest==7.4.4
black==23.12.1
flake8==7.0.0

# Additional dependencies for audio processing
ffmpeg-python==0.2.0
